Get rid of not needed struct wrapping
The struct nl_handle was wrapped inside struct _nlconnection.  This
is really not needed if open_netlink() and close_netlink() functions
uses "pointer's pointer" (struct nl_handle **) instead.  Removes also
the need to declare a static struct _nlconnection, as the
global nlconnection variable can now be a pointer as well.

Signed-off-by: David Sommerseth <davids@redhat.com>