Fix bad loop condition within get_devices()
get_devices() includes this loop:

     char buffer[256];

     ...snip...

     char *end = buffer;

     ...snip...

     while (end && *end != ':')
         end++;

The condition "end" within the while guard will always be true, given that
buffer is on the stack and thus will never be near the zero value.  It
should be *end instead, to check for the NUL terminator byte: this code
will likely crash if the string does not contain a colon character.

This appears to have been present in the initial commit of the code
(8d6ad996f5d60d569532cdba4febb19c69bdf488)

Found by Braňo Náter using the "cppcheck" static analyzer.