Fix memory leaks in get_interfaces_info()
The first half of get_interfaces_info() potentially allocates fetch_devs
using calloc, setting fetch_devs_len to a value which may or may not be
non-zero.

In particular, given a tuple argument containing all non-strings,
allocation occurs, but fetch_devs_len is set to zero, so it's not correct
to use (fetch_devs_len > 0) as the condition for freeing the memory on the
primary exit path, as this would leak fetch_devs (introduced in
bfdcac6b16806416a6c0295fcfad5d820595d88c)

There are also two error-handling paths that fail to free it (introduced in
4f0295fca2cfd933f4b9b539d5505cb24e4d420c)

Instead of this logic, simply initialize it to NULL, and pass it to free on
every exit route of the second half of the function: free(NULL) is
guaranteed to be a no-op.

Found by Braňo Náter using the "cppcheck" static analyzer.